[AN] refactor: DomainException 및 Mapper 구조 변경 및 메시지 추가 #742
*   [AN] refactor: DomainException을 단일 파일로 이동 및 기본 메시지 생성자 추가
*   [AN] refactor: DomainExceptionMapper를 DomainException 파일 내부로 이동
*   [AN] refactor: 예외 상황별 오류 메시지 상수 정의 및 매핑 로직 구체화
