--- android/app/src/main/java/com/onair/hearit/domain/exception/DomainException.kt
@@ -1,7 +1,0 @@
-package com.onair.hearit.domain.exception
-
-sealed class DomainException : Exception() {
-    data object NetworkConnection : DomainException()
-
-    data object UserNotRegistered : DomainException()
-}
